% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clock}
\alias{clock}
\title{Plot clock}
\usage{
clock(
  x = format(Sys.time(), "\%H:\%M"),
  time_label = as.roman(1:12),
  rotate_text = FALSE,
  text_color = "black",
  bg_color = "white",
  pointer_color = "black"
)
}
\arguments{
\item{x}{time, defalut: format(Sys.time(), "\%H:\%M"), e.g. 12:30}

\item{time_label}{time_label, default: as.roman(1:12)}

\item{rotate_text}{rotate_text, FALSE}

\item{text_color}{text_color, "black"}

\item{bg_color}{bg_color, "white"}

\item{pointer_color}{pointer_color, "black"}
}
\value{
ggplot
}
\description{
Plot clock
}
\examples{
clock()
}
\references{
\url{https://allancameron.github.io/geomtextpath/}
}
