% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{convert_chr_to_matrix}
\alias{convert_chr_to_matrix}
\title{convert a character to 01 matrix}
\usage{
convert_chr_to_matrix(
  char,
  size = 32,
  font_file = NULL,
  picture_dir = tempdir()
)
}
\arguments{
\item{char}{a character}

\item{size}{32}

\item{font_file}{font_file}

\item{picture_dir}{where to save the temporary picture}
}
\value{
chr_mat
}
\description{
convert a character to 01 matrix
}
\examples{
convert_chr_to_matrix("A")
}
