% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{convert_img_to_matrix}
\alias{convert_img_to_matrix}
\title{convert a imgage to 01 matrix}
\usage{
convert_img_to_matrix(image_file, size = 32, breaks = 2)
}
\arguments{
\item{image_file}{image_file}

\item{size}{32}

\item{breaks}{breaks, default 2}
}
\value{
chr_mat
}
\description{
convert a imgage to 01 matrix
}
