% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{life_game}
\alias{life_game}
\title{Life Game Simulation}
\usage{
life_game(
  save_file = NULL,
  size = 20,
  time = 20,
  fps = 0.75,
  colors = c("black", "green4"),
  ...
)
}
\arguments{
\item{save_file}{gif filename}

\item{size}{size of the world}

\item{time}{how many times the life game continue.}

\item{fps}{fps, 0.75}

\item{colors}{c("green4", "black")}

\item{...}{add}
}
\value{
a gif file
}
\description{
Life Game Simulation
}
\references{
\code{https://zhuanlan.zhihu.com/p/136727731}
}
