% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_LED}
\alias{make_LED}
\title{make a LED screen}
\usage{
make_LED(
  chars = "HHH",
  save_file = NULL,
  LED_width = 64,
  speed = 2,
  colors = c("grey", "red2"),
  ...
)
}
\arguments{
\item{chars}{chars}

\item{save_file}{save_file}

\item{LED_width}{LED_width, 64}

\item{speed}{2}

\item{colors}{c("grey","red2")}

\item{...}{add}
}
\value{
gif file
}
\description{
make a LED screen
}
