% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Moon_cake}
\alias{Moon_cake}
\title{Draw a Moon_cake (Mid- Autumn Festival) using ggplot2}
\usage{
Moon_cake(cake_color = "orange", cake_fill = "white", stuffing = "五仁")
}
\arguments{
\item{cake_color}{Color of the cake, default is "orange"}

\item{cake_fill}{Fill color of the cake, default is "white"}

\item{stuffing}{Stuffing of the cake}
}
\value{
ggplot
}
\description{
Draw a Moon_cake (Mid- Autumn Festival) using ggplot2
}
