% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{my_wordcloud}
\alias{my_wordcloud}
\title{Word cloud plot}
\usage{
my_wordcloud(
  str_vector,
  ignore_words = "Unclassified|uncultured|Ambiguous|Unknown|unknown|metagenome|Unassig",
  topN = 50
)
}
\arguments{
\item{str_vector}{string vector}

\item{ignore_words}{ignore_words}

\item{topN}{topN, 50}
}
\value{
a htmlwidget
}
\description{
Word cloud plot
}
\examples{
\donttest{
data(otutab, package = "pcutils")
my_wordcloud(taxonomy$Genus)
}
}
