% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pixel_plot}
\alias{pixel_plot}
\title{Plot a pixel image}
\usage{
pixel_plot(image, size = 32, geom = c("point", "tile"), geom_param = list())
}
\arguments{
\item{image}{image path}

\item{size}{plot size, default: 32}

\item{geom}{point or tile, default: "point"}

\item{geom_param}{geom_param, default: list(size = 4)}
}
\value{
ggplot
}
\description{
Plot a pixel image
}
