% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.chr_mat}
\alias{plot.chr_mat}
\title{Plot a chr_mat}
\usage{
\method{plot}{chr_mat}(x, colors = c("grey", "red2"), random = FALSE, ...)
}
\arguments{
\item{x}{chr_mat object}

\item{colors}{c("grey","red2")}

\item{random}{add random}

\item{...}{add}
}
\value{
plot
}
\description{
Plot a chr_mat
}
