% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot4fun-package.R
\docType{package}
\name{plot4fun-package}
\alias{plot4fun}
\alias{plot4fun-package}
\title{plot4fun: Just plot for fun}
\description{
This is a package containing many interseting functions for plotting.
}
\author{
\strong{Maintainer}: Chen Peng \email{pengchen2001@zju.edu.cn} (\href{https://orcid.org/0000-0002-9449-7606}{ORCID})

}
\keyword{internal}
