% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{show_github_calendar}
\alias{show_github_calendar}
\title{Plot a github style calendar}
\usage{
show_github_calendar(usr = "asa12138", color = NULL, save_file = NULL, ...)
}
\arguments{
\item{usr}{github username}

\item{color}{color, NULL}

\item{save_file}{save_file, NULL}

\item{...}{add}
}
\value{
a svg file
}
\description{
Plot a github style calendar
}
