% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_LED}
\alias{make_LED}
\title{make a LED screen}
\usage{
make_LED(
  chars = "SOS!",
  save_file = NULL,
  LED_width = NULL,
  speed = 32,
  fps = 10,
  colors = c("grey", "red2"),
  LED_height = 32,
  image_scale = 10,
  ...
)
}
\arguments{
\item{chars}{chars}

\item{save_file}{save_file}

\item{LED_width}{LED_width}

\item{speed}{pixel speed, default 32}

\item{fps}{frame per second, 10}

\item{colors}{c("grey","red2")}

\item{LED_height}{LED_height, 64}

\item{image_scale}{image scale, 10}

\item{...}{add}
}
\value{
gif file
}
\description{
make a LED screen
}
\examples{
\donttest{
make_LED()
}
}
