% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot4fun-package.R
\docType{package}
\name{plot4fun-package}
\alias{plot4fun}
\alias{plot4fun-package}
\title{plot4fun: Just Plot for Fun}
\description{
Explore the world of R graphics with fun and interesting plot functions! Use make_LED() to create dynamic LED screens, draw interconnected rings with Olympic_rings(), and make festive Chinese couplets with chunlian(). Unleash your creativity and turn data into exciting visuals!
}
\author{
\strong{Maintainer}: Chen Peng \email{pengchen2001@zju.edu.cn} (\href{https://orcid.org/0000-0002-9449-7606}{ORCID})

}
\keyword{internal}
